+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/PairedResultGraph/Reflow/DynamicEventline.jsx
@@ -47,6 +47,8 @@ const PairedResultDynamicEventlinePanningExample = () => {
       case 'panRight':
         hour = panState.initial + panState.factor;
         break;
+      default:
+        return panState;
     }
 
     graph.config.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();