+++ packages/terra-graphs/src/components/Bar/BarGraph.jsx
@@ -41,7 +41,7 @@ const BarGraph = ({
   // initial dataset load
   React.useEffect(() => {
     if (!graph || graphLoadedRef.current) {
-      return;
+      return undefined;
     }
     const timeoutIds = [];
 
@@ -66,7 +66,6 @@ const BarGraph = ({
 
     graphLoadedRef.current = true;
 
-    // eslint-disable-next-line consistent-return
     return () => {
       timeoutIds.forEach((id) => { clearTimeout(id); });
     };